pyproject.toml
[project]
name = "rhythmix"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.42.17",
    "celery>=5.6.0",
    "essentia>=2.1b6.dev1389",
    "fastapi[standard]>=0.128.0",
    "ffmpeg-python>=0.2.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.21",
]
